# Conflicts:
#	data_manager_sql.py

+++ server.py
@@ -46,46 +46,57 @@ def display_question(question_id):
     tags = data_manager_sql.get_tags(question_id)
     for answer in answers_data:
         comment_data[answer['id']] = data_manager_sql.get_comments_by_answer_id(answer['id'])
-    return render_template("question_page.html", question_data=question_data, answers=answers_data, comments=comment_data, tags=tags)
+    return render_template("question_page.html", question_data=question_data, answers=answers_data,
+                           comments=comment_data, tags=tags)
 
 
 @app.route("/add-question", methods=['GET', 'POST'])
 def add_question():
-    user_id = session['id']
-    if request.method == 'POST':
-        util.upload_image(request.files['image'])
-        question_id = data_manager_sql.add_question(request.form['title'], request.form['message'], user_id, request.files['image'].filename)['id']
-        return redirect(f'/question/{question_id}')
-    return render_template("add-edit-question.html")
+    if 'id' in session:
+        if request.method == 'POST':
+            user_id = session['id']
+            util.upload_image(request.files['image'])
+            question_id = data_manager_sql.add_question(request.form['title'], request.form['message'], user_id, request.files['image'].filename)['id']
+            return redirect(f'/question/{question_id}')
+        return render_template("add-edit-question.html")
+    else:
+        return redirect("/powerpuff_warning")
 
 
 @app.route("/question/<question_id>/edit", methods=['GET', 'POST'])
 def edit_question(question_id):
-    if request.method == 'POST':
-        user_id = data_manager_sql.get_question_by_id(question_id)['user_id']
-        if session['id'] == user_id:
+    user_id = data_manager_sql.get_question_by_id(question_id)['user_id']
+    if 'id' in session and session['id'] == user_id:
+        if request.method == 'POST':
             data_manager_sql.edit_question(question_id, request.form['title'], request.form['message'], request.files['image'].filename)
             if request.files['image']:
                 util.upload_image(request.files['image'])
             return redirect(f"/question/{question_id}")
-
-    return render_template("add-edit-question.html", question_data=data_manager_sql.get_question_by_id(question_id))
+        return render_template("add-edit-question.html", question_data=data_manager_sql.get_question_by_id(question_id))
+    return redirect("/powerpuff_warning")
 
 
 @app.route("/question/<question_id>/delete")
 def delete_question(question_id):
-    user_id = data_manager_sql.get_question_by_id(question_id)['user_id']
-    if session['id'] == user_id:
+    question_to_delete = data_manager_sql.get_question_by_id(question_id)
+    user_id = question_to_delete['user_id']
+    if 'id' in session and session['id'] == user_id:
+        if question_to_delete['image']:
+            util.delete_image(question_to_delete['image'])
         data_manager_sql.delete_question(question_id)
         data_manager_sql.delete_tag_if_not_in_question_tag()
-    return redirect('/list')
+        return redirect('/list')
+    return redirect("/powerpuff_warning")
 
 
 @app.route("/question/<question_id>/<vote>")
 def vote_question(question_id, vote):
-    data_manager_sql.change_question_vote_number(question_id, vote)
-    data_manager_sql.change_reputation_by_question(question_id, vote)
-    return redirect('/list')
+    if "username" in session:
+        data_manager_sql.change_question_vote_number(question_id, vote)
+        data_manager_sql.change_reputation_by_question(question_id, vote)
+        return redirect('/list')
+    else:
+        return redirect("/powerpuff_warning")
 
 
 @app.route("/bonus-questions")
@@ -96,38 +107,43 @@ def bonus_questions():
 @app.route("/question/<question_id>/new-answer", methods=['GET', 'POST'])
 def post_answer(question_id):
     if 'username' in session:
-        username = session['username']
-        user_id = data_manager_sql.get_user_id_by_user_name(username)['user_id']
+        user_id = session['id']
         if request.method == 'POST':
             util.upload_image(request.files['image'])
             data_manager_sql.add_new_answer(question_id, request.form['message'], user_id, request.files['image'].filename)
             return redirect(url_for('display_question', question_id=question_id))
-    return render_template("add-edit-answer.html", question=data_manager_sql.get_question_by_id(question_id),
-                           answers=data_manager_sql.get_answers(question_id), answer=False)
+        return render_template("add-edit-answer.html", question=data_manager_sql.get_question_by_id(question_id),
+                               answers=data_manager_sql.get_answers(question_id), answer=False)
+    return redirect("/powerpuff_warning")
 
 
 @app.route("/answer/<answer_id>/edit", methods=['GET', 'POST'])
 def edit_answer(answer_id):
-    if 'username' in session:
-        answer_to_edit = data_manager_sql.get_answer_by_id(answer_id)
+    answer_to_edit = data_manager_sql.get_answer_by_id(answer_id)
+    user_id = answer_to_edit['user_id']
+    if 'id' in session and session['id'] == user_id:
         question_id = answer_to_edit['question_id']
         if request.method == 'POST':
             data_manager_sql.edit_answer(answer_id, request.form['message'], request.files['image'].filename)
             if request.files['image']:
                 util.upload_image(request.files['image'])
             return redirect(url_for('display_question', question_id=question_id))
-    return render_template('add-edit-answer.html', answer=answer_to_edit, question=data_manager_sql.get_question_by_id(question_id))
+        return render_template('add-edit-answer.html', answer=answer_to_edit,
+                               question=data_manager_sql.get_question_by_id(question_id))
+    return redirect("/powerpuff_warning")
 
 
 @app.route("/answer/<answer_id>/delete")
 def delete_answer(answer_id):
-    if 'username' in session:
-        answer_to_delete = data_manager_sql.get_answer_by_id(answer_id)
+    answer_to_delete = data_manager_sql.get_answer_by_id(answer_id)
+    user_id = answer_to_delete['user_id']
+    if 'id' in session and session['id'] == user_id:
         if answer_to_delete['image']:
             util.delete_image(answer_to_delete['image'])
         data_manager_sql.delete_answer(answer_id)
         question_id = answer_to_delete['question_id']
-    return redirect(url_for('display_question', question_id=question_id))
+        return redirect(url_for('display_question', question_id=question_id))
+    return redirect("/powerpuff_warning")
 
 
 @app.route("/answer/<answer_id>/<vote>")
@@ -137,7 +153,8 @@ def vote_answer(answer_id, vote):
         data_manager_sql.change_answers_vote_number(vote, answer_id)
         data_manager_sql.change_reputation_by_answer(answer_id, vote)
         question_id = answer_to_vote['question_id']
-    return redirect(url_for('display_question', question_id=question_id))
+        return redirect(url_for('display_question', question_id=question_id))
+    return redirect("/powerpuff_warning")
 
 
 @app.route("/answer_acceptance/<answer_id>/<acceptance_value>")
@@ -150,23 +167,23 @@ def accept_refuse_answer(answer_id, acceptance_value):
 
 @app.route("/question/<question_id>/new-comment", methods=['GET', 'POST'])
 def add_comment_to_question(question_id):
-    if request.method == 'POST':
-        if 'username' in session:
-            user_id = data_manager_sql.get_user_id_by_user_name(session['username'])['user_id']
-            data_manager_sql.add_comments_to_question(question_id, request.form['message'], user_id)
+    if 'id' in session:
+        if request.method == 'POST':
+            data_manager_sql.add_comments_to_question(question_id, request.form['message'], session['id'])
             return redirect(url_for('display_question', question_id=question_id))
-    return render_template("add-edit-comment.html")
+        return render_template("add-edit-comment.html")
+    return redirect("powerpuff_warning")
 
 
 @app.route("/answer/<answer_id>/new-comment", methods=['GET', 'POST'])
 def add_comment_to_answer(answer_id):
-    if request.method == 'POST':
-        if 'username' in session:
-            user_id = data_manager_sql.get_user_id_by_user_name(session['username'])['user_id']
-            data_manager_sql.add_comments_to_answer(answer_id, request.form['message'], user_id)
+    if 'id' in session:
+        if request.method == 'POST':
+            data_manager_sql.add_comments_to_answer(answer_id, request.form['message'], session['id'])
             question_id = data_manager_sql.get_answer_by_id(answer_id)['question_id']
             return redirect(url_for('display_question', question_id=question_id))
-    return render_template("add-edit-comment.html")
+        return render_template("add-edit-comment.html")
+    return redirect("powerpuff_warning")
 
 
 @app.route("/comment/<comment_id>/edit", methods=['GET', 'POST'])
@@ -276,7 +293,7 @@ def list_users():
         users = data_manager_sql.get_users()
         return render_template("users-list.html", users=users)
     else:
-        return render_template("powerpuff_warning.html")
+        return redirect("/powerpuff_warning")
 
 
 @app.route("/user/<user_id>")
@@ -286,6 +303,11 @@ def display_user(user_id):
     answers = data_manager_sql.get_answers_by_user_id(user_id)
     comments = data_manager_sql.get_comments_by_user_id(user_id)
     return render_template("user.html", user=user, questions=questions, answers=answers, comments=comments)
+
+
+@app.route("/powerpuff_warning")
+def display_warning():
+    return render_template("powerpuff_warning.html")
 
 
 if __name__ == "__main__":
